requirements.txt
uvicorn==0.22.0
torch==1.13.0
transformers==4.27.0
sentence-transformers==2.2.0
langchain==0.0.195
typing-extensions==4.5.0
openai==0.27.0
python-dotenv==0.21.1
fastapi==0.115.6
uvicorn==0.22.0
huggingface_hub==0.11.1
langgraph==0.2.59
pydantic==2.10.3